net: fix reconnect deadlock, close() hang, and fragment handling
Four transport defects, each unrecoverable at runtime:

- Reconnect deadlock. `onConnected()` ran the reconnect handler inline
  from the I/O thread *before* `readLoop()` started. A handler doing
  what reconnect handlers exist to do — re-registering models, via
  `sendSync` — then waited on `_syncCv` for a reply only that same
  thread could deliver. No timeout, so the transport wedged forever.
  Handlers now run on a dedicated thread; requests coalesce and a
  throwing handler no longer takes the thread down.

- `close()` took a connection's write mutex before `shutdownBoth()`. A
  client thread blocked in `sendAll` against a stalled peer holds that
  mutex until the send completes, and the shutdown is exactly what
  unblocks it — so `close()`, reached from the destructor, waited
  forever on the thing it was trying to interrupt.

- `SocketServer` never opened or closed connection scopes: the
  reclamation machinery was wired into `QtWebSocketServer` only, so
  every model registered over this transport outlived its connection.

- `WsFrameReader` rejected all fragmented frames. A peer fragments
  whenever a message exceeds its outgoing frame size and Qt defaults
  that to 512 KiB, so this broke interop with the transport this
  project ships. Now reassembled, with control frames passing through
  mid-message and the reassembled total bounded.

Also retries EINTR in `accept()` (a stray signal killed the accept
loop) and fixes GCC -Werror violations across morph::net, which had
never been compiled by GCC anywhere in CI.

+++ include/morph/net/detail/ws_handshake.hpp
@@ -102,7 +102,9 @@ inline ParsedWsUrl parseWsUrl(std::string_view url) {
 inline std::string buildClientHandshakeRequest(const ParsedWsUrl& url, std::string_view key) {
     std::string req;
     req += "GET " + url.path + " HTTP/1.1\r\n";
-    req += "Host: " + url.host + ":" + std::to_string(url.port) + "\r\n";
+    // static_cast<unsigned>: see TcpSocket::connect — std::to_string has no
+    // uint16_t overload, and the promotion GCC would pick trips -Wsign-promo.
+    req += "Host: " + url.host + ":" + std::to_string(static_cast<unsigned>(url.port)) + "\r\n";
     req += "Upgrade: websocket\r\n";
     req += "Connection: Upgrade\r\n";
     req += "Sec-WebSocket-Key: " + std::string(key) + "\r\n";
